requirements.txt
openai
python-dotenv
sounddevice
numpy
pyttsx3
webrtcvad
